frame_suite::virtuals

Trait TryIntoTag

pub trait TryIntoTag<R, Discriminant: DiscriminantTag = ()> {
    type Error;

    // Required method
    fn try_into_tag(self) -> Result<R, Self::Error>;
}

Fallible version of IntoTag.

Attempts to convert self into another representation under a given discriminant, returning an error if the conversion fails.

Type Parameters

R: Target type (may be generic or non-concrete).

Discriminant: A concrete marker type used to disambiguate conversions.

Default Discriminant

Discriminant = (): uses the unit type as a default tag, meaning no additional disambiguation when a single interpretation exists.

Required Associated Types

type Error

Required Methods

fn try_into_tag(self) -> Result<R, Self::Error>